# openapi.json
{
  "x-openai-isConsequential": false,
  "x-openai-baseUrl": "https://gpt-to-sheet.onrender.com",
  "openapi": "3.1.0",
  "info": {
    "title": "GPT-to-Sheet API",
    "description": "API for logging financial transactions to Google Sheets using service account authentication.",
    "version": "2.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "https://gpt-to-sheet.onrender.com",
      "description": "Production server"
    }
  ],
  "tags": [
    {
      "name": "transactions",
      "description": "Transaction management endpoints"
    },
    {
      "name": "sheets",
      "description": "Google Sheets operations"
    }
  ],
  "paths": {
    "/api/log-data-to-sheet": {
      "post": {
        "tags": ["transactions"],
        "summary": "Log transaction to Google Sheets",
        "description": "Logs single or bulk transactions to a specified Google Sheet using service account authentication.",
        "operationId": "logDataToSheet",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["spreadsheetId", "sheetName"],
                "properties": {
                  "spreadsheetId": {
                    "type": "string",
                    "description": "The ID of the Google Sheet to write to"
                  },
                  "sheetName": {
                    "type": "string",
                    "description": "The name of the sheet within the spreadsheet"
                  },
                  "data": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/TransactionData"
                      },
                      {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "required": ["item", "amount"],
                          "properties": {
                            "item": {
                              "type": "string",
                              "description": "Item name"
                            },
                            "amount": {
                              "type": "number",
                              "description": "Item amount",
                              "format": "float"
                            }
                          }
                        }
                      }
                    ]
                  },
                  "meta": {
                    "$ref": "#/components/schemas/TransactionData"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged the transaction(s)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to write to sheet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/get-sheet-data": {
      "post": {
        "tags": ["sheets"],
        "summary": "Get data from Google Sheets",
        "description": "Retrieves data from a specified Google Sheet using service account authentication.",
        "operationId": "getSheetData",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["spreadsheetId", "sheetName"],
                "properties": {
                  "spreadsheetId": {
                    "type": "string",
                    "description": "The ID of the Google Sheet to read from"
                  },
                  "sheetName": {
                    "type": "string",
                    "description": "The name of the sheet within the spreadsheet"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully retrieved sheet data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDataResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to read from sheet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/service-account": {
      "get": {
        "tags": ["sheets"],
        "summary": "Get service account email",
        "description": "Returns the email address of the service account being used.",
        "operationId": "getServiceAccount",
        "responses": {
          "200": {
            "description": "Successfully retrieved service account email",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "serviceAccount": {
                      "type": "string",
                      "description": "Service account email address"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to get service account email",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TransactionData": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Transaction date (YYYY-MM-DD)",
            "format": "date"
          },
          "time": {
            "type": "string",
            "description": "Transaction time (HH:mm:ss)",
            "format": "time"
          },
          "accountName": {
            "type": "string",
            "description": "Account name"
          },
          "transactionType": {
            "type": "string",
            "description": "Type of transaction"
          },
          "category": {
            "type": "string",
            "description": "Transaction category"
          },
          "allowances": {
            "type": "string",
            "description": "Allowances"
          },
          "deductions": {
            "type": "string",
            "description": "Deductions"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount",
            "format": "float"
          },
          "establishment": {
            "type": "string",
            "description": "Establishment name"
          },
          "receiptNumber": {
            "type": "string",
            "description": "Receipt number"
          },
          "paymentMethod": {
            "type": "string",
            "description": "Method of payment"
          },
          "cardUsed": {
            "type": "string",
            "description": "Card used for payment"
          },
          "linkedBudgetCategory": {
            "type": "string",
            "description": "Linked budget category"
          },
          "onlineTransactionId": {
            "type": "string",
            "description": "Online transaction ID"
          },
          "mappedOnlineVendor": {
            "type": "string",
            "description": "Mapped online vendor"
          },
          "reimbursable": {
            "type": "string",
            "description": "Whether transaction is reimbursable"
          },
          "reimbursementStatus": {
            "type": "string",
            "description": "Status of reimbursement"
          },
          "interestType": {
            "type": "string",
            "description": "Type of interest"
          },
          "taxWithheld": {
            "type": "number",
            "description": "Tax withheld amount",
            "format": "float"
          },
          "taxDeductible": {
            "type": "string",
            "description": "Whether transaction is tax deductible"
          },
          "taxCategory": {
            "type": "string",
            "description": "Tax category"
          },
          "bankIdentifier": {
            "type": "string",
            "description": "Bank identifier"
          },
          "transactionMethod": {
            "type": "string",
            "description": "Method of transaction"
          },
          "transferMethod": {
            "type": "string",
            "description": "Method of transfer"
          },
          "referenceId": {
            "type": "string",
            "description": "Reference ID"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "processed": {
            "type": "string",
            "description": "Processing status"
          }
        }
      },
      "LogResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "message": {
            "type": "string",
            "description": "Success or error message"
          },
          "transactionId": {
            "type": "string",
            "description": "Generated transaction ID (format: TXN-XXXX)"
          },
          "receiptId": {
            "type": "string",
            "description": "Generated receipt ID (format: REC-XXXX)"
          },
          "results": {
            "type": "object",
            "properties": {
              "methods": {
                "type": "object",
                "properties": {
                  "serviceAccount": {
                    "type": "boolean"
                  },
                  "oauth": {
                    "type": "boolean"
                  },
                  "queue": {
                    "type": "boolean"
                  }
                }
              },
              "primaryMethod": {
                "type": "string"
              },
              "success": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "GetDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "2D array of cell values with 29 columns",
            "items": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "minItems": 29,
              "maxItems": 29
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "results": {
            "type": "object",
            "properties": {
              "methods": {
                "type": "object",
                "properties": {
                  "serviceAccount": {
                    "type": "boolean"
                  },
                  "oauth": {
                    "type": "boolean"
                  },
                  "queue": {
                    "type": "boolean"
                  }
                }
              },
              "primaryMethod": {
                "type": "string"
              },
              "success": {
                "type": "boolean"
              }
            }
          }
        }
      }
    }
  }
} 